added validation and verification of session in admin section

+++ admin/delete.php
@@ -1,20 +1,29 @@
 <?php
-include("_con.php");
 
+session_start();
+if (isset($_SESSION['username'])) {
 
-
-$del = $_GET['dl'];
-
-$query = "DELETE FROM contact WHERE sl = ?";
-$stmt = $conn->prepare($query);
-$stmt->bind_param("i", $del);
-
-if ($stmt->execute()) {
-    header("Location: index.php");
-} else {
-    echo "Error deleting record: " . $stmt->error;
+    include("_con.php");
+    $del = $_GET['dl'];
+    
+    $query = "DELETE FROM contact WHERE sl = ?";
+    $stmt = $conn->prepare($query);
+    $stmt->bind_param("i", $del);
+    
+    if ($stmt->execute()) {
+        header("Location: index.php");
+    } else {
+        echo "Error deleting record: " . $stmt->error;
+    }
+    
+    $stmt->close();
+    $conn->close();
+}
+else{
+    header("Location: adminLogin.php");
+    exit;
 }
 
-$stmt->close();
-$conn->close();
+
+
 ?>
